desktop/requirements.txt
PyNaCl>=1.5.0
cryptography>=41.0.0
requests>=2.31.0
Pillow>=10.0.0
pystray>=0.19.5
qrcode>=7.4
